support floating-point bonus values

+++ Common/src/main/java/com/tristankechlo/toolleveling/client/ToolLevelingTableScreen.java
@@ -95,9 +95,9 @@ public class ToolLevelingTableScreen extends AbstractContainerScreen<ToolLevelin
             this.successChanceField.setLines(List.of(ComponentUtil.TITLE_SUCCESS_CHANCE, chanceText, minChanceText, maxChanceText));
 
             // update bonus items
-            int iterations = this.getMenu().getCycles();
-            int minStrength = this.getMenu().getMinLevels();
-            int strength = this.getMenu().getLevels();
+            float iterations = this.getMenu().getCycles();
+            float minStrength = this.getMenu().getMinLevels();
+            float strength = this.getMenu().getLevels();
             Component summary = ComponentUtil.makeSummary(".summary.field_text", iterations, minStrength, strength);
             int maxPossibleWidth = this.width - (this.leftPos + this.imageWidth) - 10;
             this.bonusItemField.setLines(List.of(ComponentUtil.TITLE_BONUSES, summary), font, maxPossibleWidth);
